Fix bug in rpc-reply errors parsing.
There was an error when parsing error-tag content.
Allowed content was not correctly identified as valid, causing the parser to fail. 
The error was caused by an attempt to transform error-tag content into a constant name defined in ErrorTag class. This happened in TransportContentparser (lines 265-269).

This patch fixes explained bug by looking up desired constant given error-tag value, instead of trying to transform the second into the first.
This approach is also applied to error-type and error-severity parsing.

Thanks Diego Carucci for reporting this bug.