Add IQuery interface and load-configuration support
Query.toXML was getting a little out of hand as it was becoming
the union of all possible messages. I've introduced an IQuery type
to allow for multiple implementations.

The existing Query class is still there but for load-configuration
support I've added a new type. It also constructs the XML through
the DOM API to make it a little more resilient.